requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
idna==2.10
inflection==0.5.1
more-itertools==8.5.0
multidict==4.7.6
multitasking==0.0.9
numpy==1.19.1
pandas==1.1.1
python-dateutil==2.8.1
pytz==2020.1
Quandl==3.5.2
requests==2.24.0
six==1.15.0
urllib3==1.25.10
websockets==8.1
yarl==1.4.2
yfinance==0.1.54
